Bug 1874684 - Part 8: Prefer integer types for TimeDuration. r=mgaudet
The previous changes make it possible to switch most `TimeDuration` fields to
use `int64_t` instead of `double`.

Differential Revision: https://phabricator.services.mozilla.com/D198542

+++ js/src/builtin/temporal/Calendar.cpp
@@ -3086,7 +3086,7 @@ static bool BuiltinCalendarAdd(JSContext* cx, const PlainDate& date,
       duration.years,
       duration.months,
       duration.weeks,
-      duration.days + balanceResult.days,
+      duration.days + double(balanceResult.days),
   };
   return AddISODate(cx, date, addDuration, overflow, result);
 }
